chore: add funcrion create archive into client page
